// types.ts
export enum Environment {
  Production = 1,
  ProductionRestricted = 2, // Test environment
}

export enum InscriptionType {
  CNPJ = 1,
  CPF = 2,
}
// ...
export interface S1000Data {
  tpAmb: Environment | number | string;
  procEmi: number | string;
  verProc: string;
  tpInsc: InscriptionType | number | string;
  nrInsc: string;
  dtIni: string;
  iniValid?: string;
  acao: string;
  classTrib: string;
  indCoop: number | string;
  indConstr: number | string;
  indDesFolha: number | string;
  nmCtt: string;
  cpfCtt: string;
  email: string;
  foneFixo?: string;
}